Change assigning a priority to a new task
Earlier a priority was assigned as the ui project.tasks.length,
considering it is zero based. But in such a way there will definitely
be inconsistences if some of the project tasks has been previously
deleted. Say, we have tasks with priorities 0, 1, 2, 3, 4; then the one
with 2 is deleted, so 0, 1, 3, 4 left; if later we add another one with
a priority based on the array length, it'll be 4 with duplication: 0,
1, 3, 4, 4. So, need to get the project's task with max priority and add 1.